+++ Scorer/Dto/User.cs
@@ -1,0 +1,4 @@
+﻿namespace Scorer.Dto
+{
+    internal sealed record class User(string Username, string HtmlUrl);
+}
