+++ Scorer/Services/IApiFacade.cs
@@ -1,0 +1,9 @@
+﻿using Scorer.Dto;
+
+namespace Scorer.Services
+{
+    internal interface IApiFacade
+    {
+        User GetUser(string username);
+    }
+}
