+++ Scorer/Services/IMetricsService.cs
@@ -1,4 +1,6 @@
-﻿namespace Scorer
+﻿using Scorer.Dto;
+
+namespace Scorer.Services
 {
     public interface IMetricsService
     {
